PET-131 Create: Creation Complete Button Binding
