+++ .idea/misc.xml
@@ -25,6 +25,7 @@
         <entry key="app/src/main/res/layout/activity_register_profile.xml" value="0.37604166666666666" />
         <entry key="app/src/main/res/layout/activity_select_family_type.xml" value="0.30867892261650276" />
         <entry key="app/src/main/res/layout/activity_splash.xml" value="0.25" />
+        <entry key="app/src/main/res/layout/activity_start.xml" value="0.2736212056434374" />
         <entry key="app/src/main/res/layout/activity_test.xml" value="0.37604166666666666" />
         <entry key="app/src/main/res/layout/dialog_search_breed.xml" value="0.30782385634886705" />
         <entry key="app/src/main/res/layout/item_breed.xml" value="0.30782385634886705" />
