+++ app/src/main/java/org/retriever/dailypet/StartActivity.kt
@@ -1,0 +1,25 @@
+package org.retriever.dailypet
+
+import android.content.Intent
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import android.widget.Button
+
+class StartActivity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_start)
+
+        val loginButton = findViewById<Button>(R.id.btn_login)
+        val testButton = findViewById<Button>(R.id.btn_test)
+
+        loginButton.setOnClickListener {
+            val nextIntent = Intent(this, LoginActivity::class.java)
+            startActivity(nextIntent)
+        }
+        testButton.setOnClickListener {
+            val nextIntent = Intent(this, TestActivity::class.java)
+            startActivity(nextIntent)
+        }
+    }
+}
