+++ app/src/main/res/values/strings.xml
@@ -1,4 +1,4 @@
 <resources>
     <string name="app_name">반려하루</string>
-    <string name="invite_comment">다른사람을 초대해서 반려동물을 함께 케어해보세요!</string>
+    <string name="invite_comment">가족 또는 친구를 초대해서 반려동물을 함께 케어해보세요!</string>
 </resources>
